chore command and log
