--- backend/.env
@@ -1,3 +1,0 @@
-...
-JWT_SECRET=verycomplexsupersecret
-JWT_REFRESH_SECRET=veryrefreshepicsecret
