+++ backend/lib/core/debug_log.dart
@@ -1,0 +1,9 @@
+import 'dart:io';
+
+import 'k_debug_mode.dart';
+
+void debugLog(String message) {
+  if (kDebugMode) {
+    stdout.writeln(message);
+  }
+}
