+++ backend/lib/core/k_debug_mode.dart
@@ -1,0 +1,2 @@
+// создать переменную , в dart app , такой переменной нет но ее можно воссоздать
+const bool kDebugMode = bool.fromEnvironment('DEBUG_MODE', defaultValue: false);
