+++ backend/lib/session/session_datasource.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
-import 'dart:io';
 
+import 'package:backend/core/debug_log.dart';
 import 'package:backend/core/log_colors.dart';
 import 'package:backend/db_client/db_client.dart';
 import 'package:backend/session/session.dart';
@@ -62,7 +62,7 @@ class SessionSqliteDatasourceImpl implements SessionDatasource {
   FutureOr<Session?> getSession({String? token, String? refreshToken, int? userId}) async {
     final entry = await _dao.getSession(refreshToken: refreshToken, token: token, userId: userId);
     if (entry == null) {
-      stdout.writeln('$red getSession: entry is null $reset');
+      debugLog('$red getSession: entry is null $reset');
       return null;
     }
     ;
