+++ backend/lib/user/user_datasource.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:backend/core/debug_log.dart';
 import 'package:backend/core/log_colors.dart';
 import 'package:backend/db_client/db_client.dart' show UserTableCompanion;
 import 'package:drift/drift.dart';
@@ -21,7 +22,7 @@ class UserDataSourceImpl implements UserDataSource {
   final UserDao _userDao;
   Future<User> createUser(CreateUserDto user) async {
     try {
-      stdout.writeln('$magenta createUser email ${user.email} p: ${user.password} $reset');
+      debugLog('$magenta createUser email ${user.email} p: ${user.password} $reset');
       // await _databaseConnection.connect();
       final entry = await _userDao.insert(
         UserTableCompanion(email: Value(user.email), password: Value(user.password), createdAt: Value(DateTime.now())),
@@ -41,6 +42,12 @@ class UserDataSourceImpl implements UserDataSource {
     // await _databaseConnection.connect();
     final entry = await _userDao.getUser(userId: userId, email: email);
     if (entry == null) return null;
-    return User(userId: entry.id, email: entry.email, createdAt: entry.createdAt, role: entry.role);
+    return User(
+      userId: entry.id,
+      email: entry.email,
+      password: entry.password,
+      createdAt: entry.createdAt,
+      role: entry.role,
+    );
   }
 }
