+++ backend/lib/web_socket/command/_ws_command.dart
@@ -1,0 +1,8 @@
+import 'dart:async';
+
+import 'package:dart_frog/dart_frog.dart';
+import 'package:dart_frog_web_socket/dart_frog_web_socket.dart';
+
+abstract class WsCommand {
+  FutureOr<void> execute(RequestContext context, String roomId, WebSocketChannel channel, dynamic payload);
+}
