+++ backend/lib/web_socket/command/join_admin_command.dart
@@ -1,0 +1,33 @@
+import 'dart:convert';
+import 'dart:io';
+
+import 'package:backend/core/log_colors.dart';
+import 'package:backend/web_socket/broadcast.dart';
+import 'package:dart_frog/dart_frog.dart';
+import 'package:dart_frog_web_socket/dart_frog_web_socket.dart';
+import 'package:sha_red/sha_red.dart';
+
+import '_ws_command.dart';
+
+class JoinAdminCommand implements WsCommand {
+  @override
+  void execute(RequestContext context, String roomId, WebSocketChannel channel, dynamic payload) async {
+    stdout.writeln('$magenta JoinAdminCommand:start ${roomId} } $reset');
+    final broadcast = context.read<Broadcast>();
+    // final user = context.read<User>();
+    // todo send error
+
+    stdout.writeln('$green JoinAdminCommand:hasCounter } $reset');
+    broadcast.subscribe(roomId, channel);
+    broadcast.broadcast(
+      'admin',
+      jsonEncode(
+        WsFromServer(
+          roomId: 'admin',
+          eventType: WsEventFromServer.adminInfo,
+          payload: IdPayload(broadcast.count).toJson(),
+        ).toJson(),
+      ),
+    );
+  }
+}
